expand env vars in configured slurm_binpath
Per discussion in #9, we want to restore the previous functionality
that shell vars within the configured slurm_binpath would get
expanded.

It was a bit unsettling that it was shelling out to run `echo "%s/%s"`
to get the result, first because it seemed kind of roundabout, but
also unsettlingly because in theory backtick expansions get performed
and can run arbitrary commands.  Also interestingly, a config value
that contains doublequotes can do some interesting things to get shell
glob wildcard characters to expand, add redirection characters to
clobber arbitrary files, and add semicolons to start arbitrary new
command lines.  But, depending on your point of view, that could all
be "a feature not a bug."  Haha.

Anyway, here, instead, we do a simple substitution of anything that
looks like a shell variable, namely any occurrences of $VAR and ${VAR}

No attempt to do any sort of quoting or backslash escaping is done,
because (1) that's easy to get wrong, and (2) i hope it's reasonable
to bet that people's slurm binpaths do not contain dollar signs or
backslashes.