remove extra dumb logic to "pre-remove" submit file
the submit file is created as a temp file name, with 0600 permissions.

this means that the file will always be there, and it will always be
a new file.  It should never contain anything.  (Even if it did, it
gets immediately truncated when opened for writing.)

So it's pointless to remove it.

But it's not _harmless_ to remove it.

By removing it, the next copy that gets created gets the shell's default
permissions, which depends on umask.  Perhaps it'll be 0644 or 0664.
But will it be 0600?  Hmm!

That is, if you're lucky.

Once the tempfile is created, another program can ls the directory to
find the temporary submit filename, and in a tight-loop, wait for it
to be removed.  Then lay down another file owned by the other user, or
worse, a symlink pointed to an arbitrary location.

Like i said, extra dumb.

SOFTWARE-3994